refactor(api): split paginated response contract
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
